Moved helper function for custom_viewer to glue.viewers.custom
